ENH: refactoring of structure of API classes
