+++ requirements-dev.txt
@@ -15,3 +15,4 @@ matplotlib
 numpydoc
 sphinx-copybutton
 sphinx_rtd_theme
+bluesky-httpserver
