+++ requirements.txt
@@ -1,2 +1,3 @@
 # List required packages in this file, one per line.
+bluesky-queueserver
 httpx
